+++ shared/modules/consensus_module.go
@@ -3,6 +3,7 @@ package modules
 //go:generate mockgen -source=$GOFILE -destination=./mocks/consensus_module_mock.go -aux_files=github.com/pokt-network/pocket/shared/modules=module.go
 
 import (
+	"github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/messaging"
 	"google.golang.org/protobuf/types/known/anypb"
 )
@@ -22,11 +23,10 @@ type ConsensusModule interface {
 
 	ConsensusStateSync
 	ConsensusPacemaker
+	ConsensusDebugModule
 
 	// Consensus Engine Handlers
 	HandleMessage(*anypb.Any) error
-	// TODO(gokhan): move it into a debug module
-	HandleDebugMessage(*messaging.DebugMessage) error
 	// State Sync messages Handler
 	HandleStateSyncMessage(*anypb.Any) error
 
@@ -74,4 +74,16 @@ type ConsensusPacemaker interface {
 type ConsensusStateSync interface {
 	GetNodeIdFromNodeAddress(string) (uint64, error)
 	GetNodeAddress() string
+}
+
+type ConsensusDebugModule interface {
+	HandleDebugMessage(*messaging.DebugMessage) error
+
+	SetHeight(uint64)
+	SetRound(uint64)
+	// REFACTOR: This should accept typesCons.HotstuffStep.
+	SetStep(uint8)
+	SetBlock(*types.Block)
+	SetLeaderId(uint64)
+	SetUtilityContext(UtilityContext)
 }